pyproject.toml
[project]
name = "vla-robotics-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.55.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "ultralytics>=8.3.239",
]
